调整使得链路损耗同步时会同步计算

+++ src/main/java/com/otn/service/impl/LinkTypeServiceImpl.java
@@ -84,7 +84,7 @@ class LinkTypeServiceImpl implements LinkTypeService {
     }
 
     @Override
-    public double calculateLoss(long versionId, String linkType, long length) {
+    public double calculateLoss(long versionId, String linkType, Float length) {
         double loss = 0;
         LinkTypeDTO linkTypeDTO = getLinkType(versionId, linkType);
         if (linkTypeDTO != null) {
